useSignIn hook to include all methods
