+++ app/hooks/useSignIn.ts
@@ -2,7 +2,7 @@ import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { useAuthState } from '../context/AuthenticationContext';
 // FIREBASE
-import { auth } from '../firebase/config';
+import { auth, firestore, googleAuthProvider, githubAuthProvider } from '../firebase/config';
 // TYPES
 import { AuthActionType } from '../types/context/AuthStateContextTypes';
 
@@ -27,16 +27,94 @@ export const useSignIn = () => {
         throw new Error('Something went wrong. User could not be signed in.');
       }
 
-      dispatch({ type: AuthActionType.SIGN_IN, payload: response.user });
+      await firestore.collection('users').doc(response.user.uid).set({
+        displayName: response.user.displayName,
+        email: response.user.email,
+      })
 
+      dispatch({ type: AuthActionType.SIGN_IN, payload: response.user });
+      
       if(!isCancelled){
         setIsLoading(false);
         setError(null);
-        router.push('/');
       }
+
+      router.push('/');
     }catch(error){
       if(!isCancelled){
         setIsLoading(false);
+        setError((error as Error).message);
+      }
+    }
+  }
+
+  // ---------Signin with Google---------
+  const signInWithGoogle = async () => {
+    /*
+      Since we are signing the user in with a popup there is
+      no need keeping track of the loading state.
+    */
+    try{
+      setError(null);
+      const response = await auth.signInWithPopup(googleAuthProvider);
+
+      if(!response.user){
+        throw new Error('Sign up with google failed. User could not be created.');
+      }
+
+      await response.user.updateProfile({
+        displayName: response.user.displayName
+      })
+
+      await firestore.collection('users').doc(response.user.uid).set({
+        displayName: response.user.displayName,
+        email: response.user.email,
+      })
+
+      dispatch({ type: AuthActionType.SIGN_IN_WITH_GOOGLE, payload: response.user });
+      
+      if(!isCancelled){
+        setError(null);
+      }
+      
+      router.push('/');
+    }catch(error){
+      if(!isCancelled){
+        setError((error as Error).message);
+      }
+    }
+  }
+
+  // ---------Signin with Github---------
+  const signInWithGithub = async () => {
+    // Same here, no need for loading state
+
+    try{
+      setError(null);
+      const response = await auth.signInWithPopup(githubAuthProvider);
+
+      if(!response.user){
+        throw new Error('Sign up with google failed. User could not be created.');
+      }
+
+      await response.user.updateProfile({
+        displayName: response.user.displayName
+      })
+
+      await firestore.collection('users').doc(response.user.uid).set({
+        displayName: response.user.displayName,
+        email: response.user.email,
+      })
+
+      dispatch({ type: AuthActionType.SIGN_IN_WITH_GITHUB, payload: response.user });
+      
+      if(!isCancelled){
+        setError(null);
+      }
+
+      router.push('/');
+    }catch(error){
+      if(!isCancelled){
         setError((error as Error).message);
       }
     }
@@ -46,5 +124,5 @@ export const useSignIn = () => {
     return () => setIsCancelled(true);
   }, []);
 
-  return { signIn, error, isLoading };
+  return { signIn, signInWithGoogle, signInWithGithub, error, isLoading };
 }
